// apps/web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        AnalizeBody: {
            /** Country */
            country?: string | null;
            /** Farms */
            farms: components["schemas"]["FarmPolygonDetailData"][];
            /** Maps */
            maps: number[];
        };
// ...
        GenerateImageBody: {
            /** Country */
            country?: string | null;
            /** Feature */
            feature: {
                [key: string]: unknown;
            };
            /** Mapid */
            mapId: number;
            /** Version */
            version?: number | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}